Fixed issues with new timeout scale
